# backend/openapi.yaml
openapi: 3.0.3
info:
  title: 家方云后端API文档
  description: 家方云小程序/公众号后端API接口文档
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://your-domain.com/api
    description: 生产环境
  - url: https://dev-domain.com/api
    description: 开发环境
  - url: http://localhost/api
    description: 本地环境
...
security:
  - bearerAuth: []
paths:
...
  /user/restore/pic:
    post:
      tags:
        - 用户相关
      summary: 还原回收站图片
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - pic_ids
              properties:
                pic_ids:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /user/destroy/pic:
    post:
      tags:
        - 用户相关
      summary: 删除回收站图片（彻底删除）
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - pic_ids
              properties:
                pic_ids:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /user/delete/pic:
    post:
      tags:
        - 用户相关
      summary: 删除我的照片
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - pic_ids
              properties:
                pic_ids:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /user/visit/pics:
    post:
      tags:
        - 用户相关
      summary: 获取足迹相册图片列表
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - fid
              properties:
                fid:
                  type: integer
                key:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /album/lists/pics:
    post:
      tags:
        - 相册管理
      summary: 获取相册图片列表
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - fid
              properties:
                fid:
                  type: integer
                key:
                  type: string
                link_share_str:
                  type: string
                share_uid:
                  type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /album/all/pics:
    post:
      tags:
        - 相册管理
      summary: 获取相册内所有图片列表
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - fid
              properties:
                fid:
                  type: integer
                key:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /album/set_top/pic:
    post:
      tags:
        - 相册管理
      summary: 设置图片置顶
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /album/delete/pic:
    post:
      tags:
        - 相册管理
      summary: 删除相册图片
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /album/move/pics:
    post:
      tags:
        - 相册管理
      summary: 移动图片（收藏或添加到相册）
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证，格式：Bearer {token}
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，2000表示成功
          example: 2000
        msg:
          type: string
          description: 提示信息
          example: success
        data:
          description: 返回数据
      required:
        - code
        - msg
        - data
...